@@ -12,7 +12,7 @@
     }
   ],
   "require": {
-    "php": ">=7.2",
+    "php": "^7.4",
     "ext-json": "*"
   },
   "require-dev": {
