@@ -5,30 +5,35 @@ namespace Freezemage\Config\Exporter;
 
 
 use Freezemage\Config\ConfigInterface;
+use Freezemage\Config\Feature\FilenameGenerator;
 
 
-class JsonExporter implements ExporterInterface {
+class JsonExporter implements ExporterInterface
+{
     /**
      * @var string
      */
-    protected $filename;
+    protected ?string $filename;
 
-    public function export(ConfigInterface $config): void {
+    public function export(ConfigInterface $config, FilenameGenerator $filenameGenerator): void
+    {
         $content = $config->getConfig();
         $content = json_encode($content);
 
         if (empty($this->filename)) {
-            $this->filename = md5($content) . '.json';
+            $this->filename = $filenameGenerator->generateFilename($content);
         }
 
         file_put_contents($this->filename, $content);
     }
 
-    public function setFilename(string $filename): void {
-        $this->filename = $filename;
+    public function getFilename(): ?string
+    {
+        return $this->filename;
     }
 
-    public function getFilename(): ?string {
-        return $this->filename;
+    public function setFilename(string $filename): void
+    {
+        $this->filename = $filename;
     }
 }
