@@ -5,33 +5,51 @@ namespace Freezemage\Config;
 
 
 use Freezemage\Config\Exporter\ExporterInterface;
-use Freezemage\Config\Feature\KeyChaining;
 use Freezemage\Config\Importer\ImporterInterface;
+use Freezemage\Config\Internal\ChainedValue;
+use OutOfBoundsException;
 
 
-class ImmutableConfig implements ConfigInterface {
-    /**
-     * @var ImporterInterface $importer
-     */
-    protected $importer;
-    /**
-     * @var ExporterInterface $exporter
-     */
-    protected $exporter;
-    /**
-     * @var array $config
-     */
-    protected $config;
-    /**
-     * @var bool $keyChaining;
-     */
-    protected $keyChaining;
+final class ImmutableConfig implements ConfigInterface
+{
+    private ImporterInterface $importer;
+    private ExporterInterface $exporter;
+    /** @var array<array-key, ChainedValue> */
+    private array $config;
+    private Settings $settings;
 
-    public function __construct(ImporterInterface $importer, ExporterInterface $exporter, KeyChaining $keyChaining) {
+    public function __construct(ImporterInterface $importer, ExporterInterface $exporter, Settings $settings)
+    {
         $this->importer = $importer;
         $this->exporter = $exporter;
-        $this->config = array();
-        $this->keyChaining = $keyChaining;
+        $this->settings = $settings;
+    }
+
+    public function __clone()
+    {
+        if (isset($this->config)) {
+            $this->config = array_map(fn(ChainedValue $value): ChainedValue => clone $value, $this->config);
+        }
+    }
+
+    /**
+     * Transforms a section of config into a separate config.
+     * Detached section behaves as a whole separate config, but points to original file.
+     * See {@link ImporterInterface::setFilename()} and {@link ExporterInterface::setFilename()}
+     * on how to change the import/export filename.
+     *
+     * @param string $name
+     * @return ImmutableConfig
+     */
+    public function extractSection(string $name): ImmutableConfig
+    {
+        $section = $this->get($name);
+
+        $config = clone $this;
+        $config->config = [];
+        $config->parse($section);
+
+        return $config;
     }
 
     /**
@@ -55,44 +73,115 @@ class ImmutableConfig implements ConfigInterface {
      *
      * @return mixed|null
      */
-    public function get(string $key, $defaultValue = null) {
-        $config = $this->getConfig();
+    public function get(string $key, $defaultValue = null)
+    {
+        $config = $this->load();
 
         if ($this->isKeyChainingEnabled()) {
             $parts = explode('.', $key);
 
-            do {
-                $part = array_shift($parts);
+            foreach ($parts as $index => $part) {
+                $config = array_filter(
+                    $config,
+                    fn(ChainedValue $value): bool => isset($value->chain[$index]) && $value->chain[$index] === $part
+                );
+            }
 
-                if (!array_key_exists($part, $config)) {
-                    return $defaultValue;
+            if (empty($config)) {
+                return $defaultValue;
+            }
+
+            $offset = count($parts);
+            return array_reduce(
+                $config,
+                static function (array $accumulator, ChainedValue $value) use ($offset) {
+                    $pointer = &$accumulator;
+
+                    foreach ($value->chain as $index => $element) {
+                        if ($index < $offset) {
+                            continue;
+                        }
+                        if (!isset($pointer[$element])) {
+                            $pointer[$element] = [];
+                        }
+                        $pointer = &$pointer[$element];
+                    }
+                    $pointer = $value->value;
+
+                    return $accumulator;
+                },
+                []
+            );
+        } else {
+            foreach ($config as $chainedValue) {
+                if ($chainedValue->getFullName() === $key) {
+                    return $chainedValue->value;
                 }
-
-                $config = $config[$part];
-            } while (!empty($parts));
-
-            return $config;
+            }
         }
 
-        return $config[$key] ?? $defaultValue;
+        return $defaultValue;
     }
-    
+
+    private function load(): array
+    {
+        if (isset($this->config)) {
+            return $this->config;
+        }
+
+        $this->config = [];
+
+        $filename = $this->importer->getFilename();
+        $config = !empty($filename) && is_file($filename) ? $this->importer->import() : array();
+
+        $this->parse($config);
+
+        return $this->config;
+    }
+
+    private function parse(array $config, array $chain = []): void
+    {
+        foreach ($config as $name => $value) {
+            $currentChain = [...$chain, $name];
+            if (is_array($value)) {
+                $this->parse($value, $currentChain);
+            } else {
+                $this->config[] = new ChainedValue($currentChain, $value);
+            }
+        }
+    }
+
+    public function isKeyChainingEnabled(): bool
+    {
+        return $this->settings->keyChaining;
+    }
+
     /**
-     * Transforms a section of config into a separate config.
-     * Detached section behaves as a whole separate config, but points to original file.
-     * See {@link ImporterInterface::setFilename()} and {@link ExporterInterface::setFilename()}
-     * on how to change the import/export filename.
+     * Lazy-loads and returns full configuration.
      *
-     * @param string $name
-     * @return ImmutableConfig
+     * @return array
      */
-    public function extractSection(string $name): ImmutableConfig {
-        $section = $this->get($name);
-        
-        $config = clone $this;
-        $config->config = array($name => $section);
-        
-        return $config;
+    public function getConfig(): array
+    {
+        $config = $this->load();
+
+        return array_reduce(
+            $config,
+            static function (array $accumulator, ChainedValue $value) {
+                $pointer = &$accumulator;
+
+                foreach ($value->chain as $element) {
+                    if (!isset($pointer[$element])) {
+                        $pointer[$element] = [];
+                    }
+                    $pointer = &$pointer[$element];
+                }
+                $pointer = $value->value;
+
+                return $accumulator;
+            },
+            []
+        );
     }
 
     /**
@@ -116,90 +205,74 @@ class ImmutableConfig implements ConfigInterface {
      *
      * @return ConfigInterface
      */
-    public function set($key, $value): ConfigInterface {
+    public function set(string $key, $value): ConfigInterface
+    {
         $clone = clone $this;
 
         if ($this->isKeyChainingEnabled()) {
             $parts = explode('.', $key);
-            $clone->config = array_merge_recursive(
-                $this->getConfig(),
-                $this->buildConfig($parts, $value)
-            );
+
+            $chains = $clone->load();
+            foreach ($parts as $index => $part) {
+                $chains = array_filter(
+                    $chains,
+                    fn(ChainedValue $value): bool => isset($value->chain[$index]) && $value->chain[$index] === $part
+                );
+            }
+            if (empty($chains)) {
+                $clone->config[] = new ChainedValue($parts, $value);
+            } else {
+                $chainedValue = array_shift($chains);
+                $chainedValue->value = $value;
+            }
         } else {
-            $clone->config[$key] = $value;
+            $clone->config[] = new ChainedValue([$key], $value);
         }
 
         return $clone;
     }
 
-    protected function buildConfig(array $parts, $value): array {
-        $part = array_shift($parts);
-
-        if (count($parts) > 0) {
-            return array($part => $this->buildConfig($parts, $value));
-        }
-
-        return array($part => $value);
-    }
-
-    /**
-     * Lazy-loads and returns full configuration.
-     *
-     * @return array
-     */
-    public function getConfig(): array {
-        if (!empty($this->config)) {
-            return $this->config;
-        }
-
-        $filename = $this->importer->getFilename();
-        $this->config = !empty($filename) && is_file($filename) ? $this->importer->import() : array();
-
-        return $this->config;
-    }
-
     /**
      * Creates new configuration.
      */
-    public function save(): void {
-        $this->exporter->export($this);
+    public function save(): void
+    {
+        $this->exporter->export($this, $this->settings->filenameGenerator);
     }
 
-    /**
-     * @param ImporterInterface $importer
-     *
-     * @codeCoverageIgnore
-     */
-    public function setImporter(ImporterInterface $importer): void {
-        $this->importer = $importer;
-    }
-
-    /**
-     * @param ExporterInterface $exporter
-     *
-     * @codeCoverageIgnore
-     */
-    public function setExporter(ExporterInterface $exporter): void {
-        $this->exporter = $exporter;
-    }
-
-    public function getImporter(): ImporterInterface {
+    public function getImporter(): ImporterInterface
+    {
         return $this->importer;
     }
 
-    public function getExporter(): ExporterInterface {
+    /**
+     * @codeCoverageIgnore
+     */
+    public function setImporter(ImporterInterface $importer): void
+    {
+        $this->importer = $importer;
+    }
+
+    public function getExporter(): ExporterInterface
+    {
         return $this->exporter;
     }
 
-    public function enableKeyChaining(): void {
-        $this->keyChaining->enable();
+    /**
+     * @codeCoverageIgnore
+     */
+    public function setExporter(ExporterInterface $exporter): void
+    {
+        $this->exporter = $exporter;
     }
 
-    public function disableKeyChaining(): void {
-        $this->keyChaining->disable();
+    public function enableKeyChaining(): void
+    {
+        $this->settings->keyChaining = true;
     }
 
-    public function isKeyChainingEnabled(): bool {
-        return $this->keyChaining->isEnabled();
+    public function disableKeyChaining(): void
+    {
+        $this->settings->keyChaining = false;
     }
 }
