@@ -7,4 +7,6 @@ namespace Freezemage\Config\Exception;
 use Throwable;
 
 
-interface ConfigException extends Throwable {}
+interface ConfigException extends Throwable
+{
+}
