@@ -7,4 +7,6 @@ namespace Freezemage\Config\Exception;
 use Exception;
 
 
-class MalformedConfigException extends Exception implements ConfigException {}
+class MalformedConfigException extends Exception implements ConfigException
+{
+}
