@@ -7,4 +7,6 @@ namespace Freezemage\Config\Exception;
 use Exception;
 
 
-class MissingConfigNameException extends Exception implements ConfigException {}
+class MissingConfigNameException extends Exception implements ConfigException
+{
+}
