@@ -7,4 +7,6 @@ namespace Freezemage\Config\Exception;
 use Exception;
 
 
-class UnsupportedFileExtensionException extends Exception implements ConfigException {}
+class UnsupportedFileExtensionException extends Exception implements ConfigException
+{
+}
