@@ -7,4 +7,6 @@ namespace Freezemage\Config\Exception;
 use OutOfBoundsException;
 
 
-class UnsupportedNestingException extends OutOfBoundsException implements ConfigException {}
+class UnsupportedNestingException extends OutOfBoundsException implements ConfigException
+{
+}
