@@ -4,7 +4,8 @@
 namespace Freezemage\Config\Importer;
 
 
-interface ImporterInterface {
+interface ImporterInterface
+{
     public function import(): array;
 
     public function setFilename(string $filename): void;
